re-structures project
