+++ lib/http_clients/blank_http_client.ts
@@ -4,14 +4,8 @@
   @describe Manages authentication and requests to APIs with no configuration
 */
 
-import { HTTPRequest, HTTPResponse, HTTPClient } from './interface'
+import { HTTPRequest, HTTPResponse, HTTPClient, HTTPClientConfig } from './interface'
 import HTTPClientAdapter from './client_adapters/interface'
-
-/**
-  BlankClientConfig
-  @describe The required information for establishing HTTP connections to Blank APIs
-*/
-export interface BlankClientConfig { timeout: number }
 
 export class BlankHTTPClient implements HTTPClient {
   client: HTTPClientAdapter
@@ -19,10 +13,10 @@ export class BlankHTTPClient implements HTTPClient {
   /**
     Initializes Blank HTTP connection and authentication information
     @constructor
-    @param {BlankClientConfig} config - The configuration information used to initialize the Blank client
+    @param {HTTPClientConfig} config - The configuration information used to initialize the Blank client
     @param {HTTPClientAdapter} httpClient - The HTTP client that will facilitate HTTP requests (e.g. axios, https, etc)
   */
-  constructor(config: BlankClientConfig, httpClient: HTTPClientAdapter) {
+  constructor(config: HTTPClientConfig, httpClient: HTTPClientAdapter) {
     this.client = httpClient
 
     this.client.Configure({
@@ -31,8 +25,8 @@ export class BlankHTTPClient implements HTTPClient {
         common: {
           "Cache-Control": "no-cache, no-store, must-revalidate",
           "Content-Type": "application/json",
-          Pragma: "no-cache",
-          Accept: "application/json",
+          "Pragma": "no-cache",
+          "Accept": "application/json",
         }
       }
     })
@@ -42,17 +36,18 @@ export class BlankHTTPClient implements HTTPClient {
     Executes the HTTP Request
     @async
     @param {HTTPRequest} request - The request assembled by the using class passed to HTTP client configured for Blank
-    @returns {Promise<object>} - The resulting data from the HTTP lookup
+    @returns {Promise<HTTPResponse>} - The resulting data from the HTTP lookup
   */
   Do = async (request: HTTPRequest): Promise<HTTPResponse> | never => {
     try {
       if(request.bearerToken)
-        request.headers = { 'Authorization': `Bearer ${request.bearerToken}` }
+        request.headers = { ...request.headers, 'Authorization': `Bearer ${request.bearerToken}` }
 
-      let resourceResponse = await this.client.Do(request)
-      let { data, headers, status, statusText } = resourceResponse
+      let { data, headers, status, statusText } = await this.client.Do(request)
+
       if( !status )
         throw new Error ( "no response returned from remote" )
+
       return { data, headers, status, statusText }
     } catch( err ) {
       err.message = "\nUnable to carry out request " + err.message
