+++ tests/main.test.ts
@@ -1,4 +1,5 @@
 import { HTTPClient, HTTPRequest, HTTPResponse } from '../lib/http_clients/interface'
+import { RepositoryEntity, HTTPRepositoryEntity } from '../lib/repositories/interface'
 
 import * as chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
@@ -8,22 +9,22 @@ export const expect = chai.expect
 
 export class GoodClient implements HTTPClient {
   mockResponse: object
-  passes: number
+  hits: number
   maxPasses: number
 
-  constructor(mockResponse?: object, passes = 1) {
+  constructor(mockResponse?: object, hits = 1) {
     this.mockResponse = mockResponse
-    this.passes = passes
+    this.hits = hits
   }
 
   Do = async (request: HTTPRequest): Promise<HTTPResponse> => {
     let response = {
-      data: this.mockResponse || (0 < this.passes ? {name: "stuff", id: 123} : [{name: "other_stuff", id: 345}]),
+      data: this.mockResponse || (0 < this.hits ? {name: "stuff", id: 123} : [{name: "other_stuff", id: 345}]),
       status: 200,
       statusText: "ok",
-      headers: 0 < this.passes ? {"After-Cursor": "asdf"} : {}
+      headers: 0 < this.hits ? {"After-Cursor": "asdf"} : {}
     }
-    this.passes--
+    this.hits--
     return response
   }
 }
@@ -31,5 +32,47 @@ export class GoodClient implements HTTPClient {
 export class BadClient implements HTTPClient {
   Do = async (request: HTTPRequest): Promise<HTTPResponse> => {
     return {  data: {message: "uh oh!"}, status: 500, statusText: "is down", headers: {} }
+  }
+}
+
+export class GoodTestRepository {
+  mockData: object[]
+
+  constructor(mockData: object[]){
+    this.mockData = mockData
+  }
+
+  Query = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return request.data ? { data: [ this.mockData[0] ] } : { data: this.mockData }
+  }
+
+  ReadResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: this.mockData[request.id as number] }
+  }
+
+  WriteResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: {...request.data, id: 1} } // add id 1 if create or just use existing id if update
+  }
+
+  DestroyResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: {} }
+  }
+}
+
+export class BadTestRepository {
+  Query = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: [], error: "Its Borked" } // Didnt find any or an error happened
+  }
+
+  ReadResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: null, error: "Its Borked" } // no response or an error
+  }
+
+  WriteResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: null, error: "Its Borked" } // no response or an error
+  }
+
+  DestroyResource = async (request: RepositoryEntity): Promise<RepositoryEntity | HTTPRepositoryEntity> => {
+    return { data: null, error: "Its Borked" } // no response or an error
   }
 }
